+++ eskema/frictionless/monkey.py
@@ -1,4 +1,5 @@
 from .pandas_plugin import create_parser
+from .parser_jsonl import read_cell_stream_create
 from .resource import ResourcePlus
 from .stream_loader import read_byte_stream_create
 
@@ -26,8 +27,10 @@ def patch_modules():
     - Don't croak when reading streams without `name` attribute.
     """
 
+    import frictionless.formats.json.parsers
     import frictionless.formats.pandas.plugin
     import frictionless.schemes.stream.loader
 
+    frictionless.formats.json.parsers.JsonlParser.read_cell_stream_create = read_cell_stream_create
     frictionless.formats.pandas.plugin.PandasPlugin.create_parser = create_parser
     frictionless.schemes.stream.loader.StreamLoader.read_byte_stream_create = read_byte_stream_create