+++ eskema/frictionless/parser_jsonl.py
@@ -1,0 +1,14 @@
+import frictionless.formats
+
+read_cell_stream_create_original = frictionless.formats.json.parsers.JsonlParser.read_cell_stream_create
+
+
+def read_cell_stream_create(self):
+    """
+    Patch to frictionless NDJSON reader to not read the whole file.
+    """
+    cell_stream = read_cell_stream_create_original(self)
+    for count, item in enumerate(cell_stream):
+        yield item
+        if count >= self.resource.detector.sample_size:
+            break