+++ eskema/settings.py
@@ -1,10 +1,10 @@
-# How many bytes to read from input data.
 from eskema.type import ContentType
-
-PEEK_BYTES = 10000
 
 # How many lines to read from input data.
 PEEK_LINES = 100
+
+# How many bytes to read from input data.
+PEEK_BYTES = PEEK_LINES * 130
 
 # Which content types to route to the "frictionless" backend.
 FRICTIONLESS_CONTENT_TYPES = [