+++ CHANGES.rst
@@ -26,6 +26,7 @@ in progress
 - Performance: Don't open resource twice when using ``frictionless`` backend
 - Performance: Improve peeking into Parquet files
 - Performance: Configure ``PEEK_LINES = 100`` instead of 1000
+- Performance: Optimize reading from remote NDJSON files
 
 
 2023-01-xx 0.0.0